src/download_and_process_data/requirements.txt
pandas==0.25.1
boto3==1.14.22
botocore==1.17.22
mysql_connector_repackaged==0.3.1
paramiko==2.7.1
selenium==3.141.0
